The continue handler has been fixed.
The continue handler was setting to few memory.
I changed a uint32_t to long.

+++ include/define.h
@@ -10,7 +10,7 @@ typedef enum {
 
 typedef struct {
     uintptr_t addr;
-    uint32_t old_byte;
+    long old_byte;
 } breakpoint_t;
 
 typedef struct {
